backend/requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
groq
deepgram-sdk
websockets
requests
motor
bcrypt
python-jose[cryptography]
email-validator
